requirements.txt
pycryptodome==3.20.0
pyperclip==1.8.2
pystray==0.19.5
Pillow==10.2.0
argon2-cffi==23.1.0
PyQt6>=6.5.0
